Add leave modals

+++ src/pages/Leaves/Leaves.tsx
@@ -13,18 +13,16 @@ import LDTable from "components/LDTable/LDTable";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { BiChevronDown, BiDotsVerticalRounded } from "react-icons/bi";
 import { fetchData } from "services/fetchData";
-import CreateLeaveModal from "./CreateLeaveModal";
-import EditLeaveModal from "./EditLeaveModal";
-
-export type User = {
-  firstName: string;
-  lastName: string;
-};
+import NewLeaveModal from "./NewLeaveModal";
+import DeleteLeaveModal from "./DeleteLeaveModal";
+import { User } from "../Employee/Employee";
+import { NewLeaveInputs } from "./NewLeaveModal";
 
 export type Leave = {
   id: number;
-  from: string;
-  to: string;
+  startAt: Date;
+  endAt: Date;
+  noOfDays: any;
   reason: string;
   status: string;
   user: User;
@@ -36,21 +34,49 @@ enum LeaveStatus {
   DECLINED = "DECLINED",
 }
 
+function calculateLeaveDays(startDate: Date, endDate: Date) {
+  const totalHours = (endDate.getTime() - startDate.getTime()) / 1000 / 3600;
+  const remainder = totalHours % 24;
+  let remainWorkDay;
+  if (remainder === 0) {
+    remainWorkDay = 0;
+  } else if (remainder >= 9) {
+    remainWorkDay = 1;
+  } else {
+    remainWorkDay = 0.5;
+  }
+  let numWorkDays = 0;
+  while (startDate <= endDate) {
+    if (startDate.getDay() !== 0 && startDate.getDay() !== 6) {
+      numWorkDays++;
+    }
+    startDate.setDate(startDate.getDate() + 1);
+  }
+  return numWorkDays - 1 + remainWorkDay;
+}
+
 function Leaves() {
   const [leaves, setLeaves] = useState<Leave[]>([]);
-  const [selectedLeave, setSelectedLeave] = useState<Leave | null>(null);
-  const [, setIsLoading] = useState(false);
+  const [selectedLeave, setSelectedLeave] = useState<Leave>();
+  const [isLoading, setIsLoading] = useState(false);
 
   const {
     isOpen: isOpenCreate,
     onOpen: onOpenCreate,
     onClose: onCloseCreate,
-  } = useDisclosure({ id: "123" });
+  } = useDisclosure();
+
   const {
     isOpen: isOpenEdit,
     onOpen: onOpenEdit,
     onClose: onCloseEdit,
-  } = useDisclosure({ id: "asdas" });
+  } = useDisclosure();
+
+  const {
+    isOpen: isOpenDelete,
+    onOpen: onOpenDelete,
+    onClose: onCloseDelete,
+  } = useDisclosure();
 
   useEffect(() => {
     const getAllLeaves = async () => {
@@ -66,12 +92,51 @@ function Leaves() {
     getAllLeaves();
   }, []);
 
-  const updateLeave = (newLeave: Leave) => {
-    const newLeaves = leaves.map((leave) => {
-      if (leave.id === newLeave.id) return newLeave;
-      return leave;
-    });
-    setLeaves(newLeaves);
+  const createLeave = async ({ startAt, endAt, reason }: NewLeaveInputs) => {
+    setIsLoading(true);
+    try {
+      const newLeave = await fetchData(`/leaves`, {
+        method: "POST",
+        body: new URLSearchParams({
+          startAt: startAt.toISOString(),
+          endAt: endAt.toISOString(),
+          reason,
+        }),
+      });
+      setIsLoading(false);
+      setLeaves([newLeave, ...leaves]);
+      onCloseCreate();
+    } catch (error) {
+      setIsLoading(false);
+      onCloseCreate();
+    }
+  };
+
+  const updateLeave = async ({ startAt, endAt, reason }: NewLeaveInputs) => {
+    if (!selectedLeave) return;
+    setIsLoading(true);
+    try {
+      let newLeave: Leave;
+      const body = new URLSearchParams({
+        startAt: startAt.toISOString(),
+        endAt: endAt.toISOString(),
+        reason,
+      });
+      newLeave = await fetchData(`/leaves/${selectedLeave.id}`, {
+        method: "PATCH",
+        body,
+      });
+      const newLeaves = leaves.map((leave) => {
+        if (leave.id === newLeave.id) return newLeave;
+        return leave;
+      });
+      setLeaves(newLeaves);
+      setIsLoading(false);
+      onCloseEdit();
+    } catch (error) {
+      setIsLoading(false);
+      onCloseEdit();
+    }
   };
 
   const changeLeaveStatus = useCallback(
@@ -81,9 +146,6 @@ function Leaves() {
         setIsLoading(true);
         const newLeave = await fetchData(`/leaves/${leave.id}/status`, {
           method: "PATCH",
-          headers: {
-            "Content-Type": "application/x-www-form-urlencoded",
-          },
           body: new URLSearchParams({ status }),
         });
         const newLeaves = leaves.map((leave) => {
@@ -99,51 +161,45 @@ function Leaves() {
     [leaves]
   );
 
-  const deleteLeave = useCallback(
-    async (leave: Leave) => {
-      try {
-        setIsLoading(true);
-        await fetchData(`/leaves/${leave.id}`, {
-          method: "DELETE",
-        });
-        const newLeaves = leaves.filter((l) => l.id !== leave.id);
-        setIsLoading(false);
-        setLeaves(newLeaves);
-      } catch (error) {
-        setIsLoading(false);
-      }
-    },
-    [leaves]
-  );
-
-  const openEdit = useCallback(
-    (leave) => {
-      setSelectedLeave(leave);
-      onOpenEdit();
-    },
-    [onOpenEdit]
-  );
+  const deleteLeave = useCallback(async () => {
+    if (!selectedLeave) return;
+    try {
+      setIsLoading(true);
+      await fetchData(`/leaves/${selectedLeave.id}`, {
+        method: "DELETE",
+      });
+      const newLeaves = leaves.filter((l) => l.id !== selectedLeave.id);
+      setIsLoading(false);
+      onCloseDelete();
+      setLeaves(newLeaves);
+    } catch (error) {
+      setIsLoading(false);
+    }
+  }, [leaves, selectedLeave, onCloseDelete]);
 
   const data = useMemo(
     () =>
       leaves.map((leave) => {
         return {
           employee: leave.user.firstName + " " + leave.user.lastName,
-          from: new Date(leave.from).toLocaleString(undefined, {
+          startAt: new Date(leave.startAt).toLocaleString(undefined, {
             weekday: "short",
             month: "short",
             day: "numeric",
             hour: "numeric",
             minute: "numeric",
           }),
-          to: new Date(leave.to).toLocaleString(undefined, {
+          endAt: new Date(leave.endAt).toLocaleString(undefined, {
             weekday: "short",
             month: "short",
             day: "numeric",
             hour: "numeric",
             minute: "numeric",
           }),
-          noOfDays: "noOfDays",
+          noOfDays: calculateLeaveDays(
+            new Date(leave.startAt),
+            new Date(leave.endAt)
+          ),
           reason: leave.reason,
           status: (
             <Menu>
@@ -166,14 +222,28 @@ function Leaves() {
             <Menu>
               <MenuButton as={IconButton} icon={<BiDotsVerticalRounded />} />
               <MenuList>
-                <MenuItem onClick={() => openEdit(leave)}>Edit</MenuItem>
-                <MenuItem onClick={() => deleteLeave(leave)}>Delete</MenuItem>
+                <MenuItem
+                  onClick={() => {
+                    setSelectedLeave(leave);
+                    onOpenEdit();
+                  }}
+                >
+                  Edit
+                </MenuItem>
+                <MenuItem
+                  onClick={() => {
+                    setSelectedLeave(leave);
+                    onOpenDelete();
+                  }}
+                >
+                  Delete
+                </MenuItem>
               </MenuList>
             </Menu>
           ),
         };
       }),
-    [leaves, changeLeaveStatus, openEdit, deleteLeave]
+    [leaves, changeLeaveStatus, onOpenDelete, onOpenEdit]
   );
 
   const columns = useMemo(
@@ -183,12 +253,12 @@ function Leaves() {
         accessor: "employee",
       },
       {
-        Header: "From",
-        accessor: "from",
+        Header: "Start At",
+        accessor: "startAt",
       },
       {
-        Header: "To",
-        accessor: "to",
+        Header: "End At",
+        accessor: "endAt",
       },
       {
         Header: "No Of Days",
@@ -209,23 +279,30 @@ function Leaves() {
     ],
     []
   );
+
   return (
     <Box>
       <Button onClick={onOpenCreate}>Add Leave</Button>
       <LDTable data={data} columns={columns} />
-      <CreateLeaveModal
+      <NewLeaveModal
         isOpen={isOpenCreate}
         onClose={onCloseCreate}
-        onFinishSubmit={(newLeave) => setLeaves([newLeave, ...leaves])}
+        isLoading={isLoading}
+        onSubmit={createLeave}
       />
-      {selectedLeave && (
-        <EditLeaveModal
-          leave={selectedLeave}
-          isOpen={isOpenEdit}
-          onClose={onCloseEdit}
-          onFinishSubmit={(newLeave) => updateLeave(newLeave)}
-        />
-      )}
+      <NewLeaveModal
+        leave={selectedLeave}
+        isOpen={isOpenEdit}
+        onClose={onCloseEdit}
+        isLoading={isLoading}
+        onSubmit={updateLeave}
+      />
+      <DeleteLeaveModal
+        isOpen={isOpenDelete}
+        onClose={onCloseDelete}
+        isLoading={isLoading}
+        onSubmit={deleteLeave}
+      />
     </Box>
   );
 }
